+++ pkg/from/fdwca/import.go
@@ -3,7 +3,7 @@ package fdwca
 func (fd *fdwca) Import(src, out string) error {
 	var err error
 
-	err = fd.dwca.Fetch(src, fd.cfg.DataDir)
+	err = fd.dwca.Fetch(src, fd.cfg.ImportDir)
 	if err != nil {
 		return err
 	}