+++ pkg/from/ftext/import.go
@@ -8,7 +8,7 @@ func (ft *ftext) Import(src, out string) error {
 	var err error
 	err = util.PrepareFileStructure(ft.cfg)
 
-	err = ft.text.Fetch(src, ft.cfg.DataDir)
+	err = ft.text.Fetch(src, ft.cfg.ImportDir)
 	if err != nil {
 		return err
 	}