+++ pkg/from/fxsv/import.go
@@ -3,7 +3,7 @@ package fxsv
 func (fx *fxsv) Import(src, out string) error {
 	var err error
 
-	err = fx.xsv.Fetch(src, fx.cfg.DataDir)
+	err = fx.xsv.Fetch(src, fx.cfg.ImportDir)
 	if err != nil {
 		return err
 	}