+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.1] - 2025-04-15 Tue
+
+Add [#16]: add update command to convert old SFGA to current schema.
 Add [#15]: convert SFGA to CoLDP.
 
 ## [v0.2.0] - 2025-03-26 Wed
@@ -34,6 +37,7 @@ Add: initial commit
 
 This document follows [changelog guidelines]
 
+[v0.2.1]: https://github.com/sfborg/sf/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/sfborg/sf/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/sfborg/sf/compare/v0.0.0...v0.1.0
 [v0.0.0]: https://github.com/sfborg/sf/tree/v0.0.0